requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pandas
openpyxl
Pillow
ImageHash
numpy

